[BACKPORT 2025.2][PLAT-21746][dr] Pass target master addresses to XClusterDeleteOutboundReplicationGroup rpc
Summary:
This diff updates the DeleteReplicationOnSource RPC to pass in the target master addresses so during a switchover the metadata on the old secondary universe is updated properly.

Also, it avoids calling DeleteUniverseReplication RPC during switchover to make sure DeleteReplicationOnSource updates the metadata.

Original diff: c79737a54260156b8975a248ebd3a68daf1cabbf / D56310

Test Plan: Created automatic DR config with a table with sequence and made sure after a switchover a new table with sequence can be created without breaking the replication

Reviewers: #yba-api-review, amindrov, jmak, nbhatia

Reviewed By: #yba-api-review, jmak

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D56381

+++ managed/src/main/java/com/yugabyte/yw/commissioner/tasks/UniverseTaskBase.java
@@ -6226,7 +6226,8 @@ public abstract class UniverseTaskBase extends AbstractTaskBase {
       boolean keepEntry,
       boolean forceDelete,
       boolean deleteSourcePitrConfigs,
-      boolean deleteTargetPitrConfigs) {
+      boolean deleteTargetPitrConfigs,
+      boolean isSwitchover) {
 
     // If target universe is destroyed, ignore creating this subtask.
     if (xClusterConfig.getTargetUniverseUUID() != null
@@ -6239,9 +6240,14 @@ public abstract class UniverseTaskBase extends AbstractTaskBase {
               forceDelete)
           .setSubTaskGroupType(UserTaskDetails.SubTaskGroupType.DeleteXClusterReplication);
     }
-    // Delete the replication group on the target universe.
-    createDeleteReplicationTask(xClusterConfig, forceDelete)
-        .setSubTaskGroupType(UserTaskDetails.SubTaskGroupType.DeleteXClusterReplication);
+    // Do not run the following for automatic mode switchover because DeleteReplicationOnSource
+    // task will delete the replication group on the target and also updates some required metadata
+    // while the DeleteUniverseReplication RPC doesn't.
+    if (!(xClusterConfig.isAutomaticDdlMode() && isSwitchover)) {
+      // Delete the replication group on the target universe.
+      createDeleteReplicationTask(xClusterConfig, forceDelete)
+          .setSubTaskGroupType(UserTaskDetails.SubTaskGroupType.DeleteXClusterReplication);
+    }
     if (xClusterConfig.getType() == ConfigType.Db) {
       // If it's in the middle of a repair, there's no replication on source.
       if (!(xClusterConfig.isUsedForDr() && xClusterConfig.getDrConfig().isHalted())) {
@@ -6361,6 +6367,21 @@ public abstract class UniverseTaskBase extends AbstractTaskBase {
       createDeleteXClusterConfigEntryTask(xClusterConfig)
           .setSubTaskGroupType(UserTaskDetails.SubTaskGroupType.DeleteXClusterReplication);
     }
+  }
+
+  protected void createDeleteXClusterConfigSubtasks(
+      XClusterConfig xClusterConfig,
+      boolean keepEntry,
+      boolean forceDelete,
+      boolean deleteSourcePitrConfigs,
+      boolean deleteTargetPitrConfigs) {
+    createDeleteXClusterConfigSubtasks(
+        xClusterConfig,
+        keepEntry,
+        forceDelete,
+        deleteSourcePitrConfigs,
+        deleteTargetPitrConfigs,
+        false /* isSwitchover */);
   }
 
   protected SubTaskGroup createDeleteDrConfigEntryTask(DrConfig drConfig) {
